feat(rollout): execute Nitin's switch decisions SW-1/2/3 (flag ON, QA merges resolved, paygw fix canonical)
Nitin approved all four queued switches (2026-06-10). This lands everything
executable from the workstation; ROLLOUT-PACKET-2026-06-10.md carries the
exact live-server remainder for IT.

SW-2 (merge 5 QA branches) — resolved by analysis, not blind merge:
- e01 + paygw branches: 0 commits ahead (already contained in production).
- e02 (skills nav), p02 (cart_badge), sa05 (theme branding): superseded by
  the canonicalized theme/sentientia (verified present at HEAD; sa05's goal
  exceeded by batch-4 "Sentientia UX").
- SA-04 was genuinely missing -> PORTED to theme/sentientia core_renderer:
  can_reach_native_course_admin() + capability gates on the
  /course/management.php + /course/index.php catalog redirects (siteadmins/
  category managers reach native management; learners keep the LXP catalog).

SW-3 (paygw security fix) — preservation regression found + fixed: the
fail-closed verifier existed only in the orphaned moodle-enhancement/payment
copy (v..0.10) while the canonical payment/gateway/airpay (webroot lineage,
v..0.09) still shipped the enrolment bypass + md5 checksum + file-scope
require_login. Canonical tree now = fixed tree (5 files + tests/). 20 files
php -l clean; deployed to local staging; sandbox transaction test explicitly
Nitin-deferred to post-deploy.

SW-1 (one-click free enrol) — new audited idempotent CLI
local/sentientia_catalog/cli/enable_oneclick_enrol.php; EXECUTED on local:
tenant 1 already ON, tenant 177 OFF->ON, Public 77 untouched. Production
runs the same command in the deploy window (packet step 4).

+++ payment/gateway/airpay/db/upgrade.php
@@ -27,10 +27,15 @@ defined('MOODLE_INTERNAL') || die();
 /**
  * Upgrade the plugin.
  *
- * @param int $oldversion the version we are upgrading from
+ * @param float $oldversion the version we are upgrading from. MUST be float
+ *   — paygw_airpay uses decimal versions (2024100700.02 .. 2024100700.10).
+ *   The previous `int` type-hint truncated the decimal and made every
+ *   `if ($oldversion < 2024100700.0X)` block fire incorrectly on any
+ *   already-upgraded DB, triggering downgrade_exception via the savepoint
+ *   call. Fixed 2026-05-24.
  * @return bool always true
  */
-function xmldb_paygw_airpay_upgrade(int $oldversion): bool
+function xmldb_paygw_airpay_upgrade(float $oldversion): bool
 {
     global $DB;
